Modified are_you_charging.sh
I realise that I don't want to know if my battery is charging or not. What I wanted to know that if my charger or power cable is connected
to the power or not. So you upower is not a great tool for this. Instead I cat the /sys/class/power_supply/BAT0/power_now file which gives
us a value 0 if the laptop is taking direct AC power thus giving us the information we need.